requirements.txt
python-osc>=1.8.0
anthropic>=0.25.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
httpx>=0.27.0
